Getting things compiling again on moderner UEs
So most of the issues around everything not compiling correctly are currently seeming to all stem from the manipulation of string tables, and it was always super jank even when it did work.
For now this side steps the issue and uses the string tables from compiled Yarn files and is not intended to be a long term solution.
Also fixes some format static assert issues.